Extract SourceHandler from LHCInfoPer*PopConAnalyzer
